Refactored points to work with doubles instead of floats.

+++ app/src/main/java/com/alexshtf/offlinenavigator/Utils.java
@@ -40,8 +40,8 @@ public class Utils {
 
     public static Point asPoint(Location location) {
         return Point.xy(
-                (float) location.getLongitude(),
-                (float) location.getLatitude()
+                location.getLongitude(),
+                location.getLatitude()
         );
     }
